src/requirements.txt
appium-python-client==5.2.4
pytest-html==4.1.1
pytest==9.0.1
selenium==4.39.0
